Allow body for DELETE
From the rfc:

  A payload within a DELETE request message has no defined semantics;
  sending a payload body on a DELETE request might cause some existing
  implementations to reject the request.

The spec does not forbid it. There are legitimate use cases for passing a
body to DELETE request; for example, when deleting a file in a Git repo it
is considered safe to pass the blob sha (so you don't end up accidentally
deleting an updated file).

GitHub API does this:

https://developer.github.com/v3/repos/contents/#delete-a-file