Use better default TLS settings
On versions of tls pre-2.0.6, the default `Supported` had an empty list
of `supportedCiphers`. Per discussion on [their issue][1], it seems like
everyone was overriding it to `ciphersuite_default` anyway, hence the
change at version 2.0.6 to make that default. But we're not doing that
here!  Luckily, the [default for `TLSSettings`][2] does do that
override, so we can use that to get the right behaviour regardless of
tls version.

[1]: https://github.com/haskell-tls/hs-tls/pull/471
[2]: https://hackage.haskell.org/package/crypton-connection-0.4.0/docs/src/Network.Connection.Types.html#line-86